Move methods from exchange composable to price composable
