+++ apps/web/src/composables/price.ts
@@ -1,4 +1,4 @@
-import { ethers } from 'ethers'
+import { Currency } from '@casimir/types'
 
 type PriceEntry = {
     time: number;
@@ -25,8 +25,40 @@ export default function usePrice() {
 
         return rate
     }
+
+    async function getConversionRateByDate (from: Currency, to: Currency, date: string) {
+        const timestamp = Math.floor(new Date(date).getTime() / 1000)
+        const options = {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json',
+                'authorization': `Apikey ${import.meta.env.PUBLIC_CRYPTO_COMPARE_API_KEY}`
+            }
+        }
+        const response = await fetch (`https://min-api.cryptocompare.com/data/pricehistorical?fsym=${from}&tsyms=${to}&ts=${timestamp}`, options)
+        const { [from]: data } = await response.json()
+        return data[to]
+    }
+
+    function convertToWholeUnits (currency: Currency, amount: number) {
+        switch (currency) {
+            case 'BTC':
+                return amount / 100000000
+                break
+
+            case 'ETH':
+                return amount / 1000000000000000000
+                break
+        
+            default:
+                break
+        }
+    }
+
     return {
-        getExchangeRate
+        getExchangeRate,
+        getConversionRateByDate,
+        convertToWholeUnits
     }
 }
